apis/core_api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
matplotlib==3.1.1
numpy==1.20.1
pandas==1.2.2
python-jose==3.2.0
PyYAML==5.4.1
pytest
Werkzeug==0.16.0